chore: move package of `LoginRequest.java`, `LoginResponse.java`, `RegisterRequest.java`
