+++ src/main/java/gdsc/nanuming/auth/dto/response/LoginResponse.java
@@ -1,4 +1,4 @@
-package gdsc.nanuming.auth.dto;
+package gdsc.nanuming.auth.dto.response;
 
 import gdsc.nanuming.jwt.dto.JwtToken;
 import lombok.Getter;
